Tag Repository

+++ cartpro/resources/views/admin/includes/common_js/bulk_action_js_0.blade.php
@@ -52,7 +52,6 @@ $("#bulk_action").on("click",function(){
                 method: "GET",
                 data: {idsArray:idsArray,action_type:action_type},
                 success: function (data) {
-                    console.log(data);
                     if(data.success){
                         $('#bulkConfirmModal').modal('hide');
                         table.rows('.selected').deselect();
@@ -75,13 +74,13 @@ $("#bulk_action").on("click",function(){
             });
         });
 
-        {{-- @if (isset($route_name_bulk_delete)) --}}
+        @if (isset($route_name_bulk_delete))
             $("#bulkDelete").on("click",function(){
                 action_type = "delete";
                 $.ajax({
-                    url: route_name_bulk_active_inactive,
+                    url: "{{route($route_name_bulk_delete)}}",
                     method: "GET",
-                    data: {idsArray:idsArray,action_type:action_type},
+                    data: {idsArray:idsArray},
                     success: function (data) {
                         console.log(data);
                         if(data.success){
@@ -106,7 +105,7 @@ $("#bulk_action").on("click",function(){
                     }
                 });
             });
-        {{-- @endif --}}
+        @endif
 
     }
 });
